refactor: refactor `front/logics/`
로직스 폴더를 구조화.
로직스 폴더 내부의 상수 파일들을 분리.
