+++ front/src/components/globalComponents/TextInput/TextInput.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import TextInputStyle from './TextInput.style';
-import regex from '../../../logics/TextInput';
+import regex from '../../../constant/TextInput';
 
 const TextInput = ({ onChange, value, onChangeComplete }) => {
   const [input, setInput] = useState(value);
