+++ front/src/logics/wordPreview/makeEmptyArray.js
@@ -1,0 +1,5 @@
+function makeEmptyArray(length) {
+  return new Array(length).fill('_');
+}
+
+export default makeEmptyArray;
