+++ front/src/pages/Home/Home.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import LoginSection from '../../components/LoginSection/LoginSection';
-import { getAutoLogIn } from '../../logics/autoLogIn';
-import checkAutoLogIn from '../../logics/checkAutoLogIn';
+import { getAutoLogIn } from '../../logics/auth/autoLogIn';
+import checkAutoLogIn from '../../logics/auth/checkAutoLogIn';
 import Background from '../../components/globalComponents/Container/Background.style';
 
 const Home = () => {
